refactor: InputMenuState, InputMenuUiEvent 클래스 이동 및 분리
