+++ app/src/main/java/com/erica/gamsung/menu/presentation/InputMenuState.kt
@@ -1,0 +1,8 @@
+package com.erica.gamsung.menu.presentation
+
+data class InputMenuState(
+    val name: String = "",
+    val price: String = "",
+    val isNameValid: Boolean = true,
+    val isPriceValid: Boolean = true,
+)
