+++ app/src/main/java/com/erica/gamsung/menu/presentation/InputMenuUiEvent.kt
@@ -1,0 +1,19 @@
+package com.erica.gamsung.menu.presentation
+
+sealed interface InputMenuUiEvent {
+    data class NameChanged(
+        val name: String,
+    ) : InputMenuUiEvent
+
+    data class PriceChanged(
+        val price: String,
+    ) : InputMenuUiEvent
+
+    data object AddMenu : InputMenuUiEvent
+
+    data class RemoveMenu(
+        val index: Int,
+    ) : InputMenuUiEvent
+
+    data object SendMenus : InputMenuUiEvent
+}
