+++ app/src/main/java/com/erica/gamsung/menu/presentation/InputMenuViewModel.kt
@@ -99,28 +99,3 @@ class InputMenuViewModel(
         return int >= 0
     }
 }
-
-sealed interface InputMenuUiEvent {
-    data class NameChanged(
-        val name: String,
-    ) : InputMenuUiEvent
-
-    data class PriceChanged(
-        val price: String,
-    ) : InputMenuUiEvent
-
-    data object AddMenu : InputMenuUiEvent
-
-    data class RemoveMenu(
-        val index: Int,
-    ) : InputMenuUiEvent
-
-    data object SendMenus : InputMenuUiEvent
-}
-
-data class InputMenuState(
-    val name: String = "",
-    val price: String = "",
-    val isNameValid: Boolean = true,
-    val isPriceValid: Boolean = true,
-)
